pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bps-review"
version = "0.1.0"
description = "Protocol-aligned systematic review pipeline for biopsychosocial framing in chronic pain reviews."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "bibtexparser>=1.4.1",
  "jinja2>=3.1.4",
  "matplotlib>=3.8",
  "networkx>=3.3",
  "numpy>=1.26",
  "openpyxl>=3.1.5",
  "pandas>=2.2",
  "pydantic>=2.9",
  "pycountry>=24.6.1",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0",
  "rapidfuzz>=3.10.0",
  "requests>=2.32",
  "rispy>=0.9.0",
  "scikit-learn>=1.5.2",
  "scipy>=1.14.1",
  "seaborn>=0.13.2",
  "typer>=0.12.5",
]

[project.scripts]
bps-review = "bps_review.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
